requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# AI and NLP
openai==1.3.0
langchain==0.0.335
transformers==4.35.2
torch==2.1.0

# Data processing
pandas==2.1.1
requests==2.31.0
beautifulsoup4==4.12.2

# Audio processing
pyttsx3==2.90
SpeechRecognition==3.10.0
librosa==0.10.1
pydub==0.25.1

# Web and API
streamlit==1.28.1
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
